Address PR-1 review: rename train.py, relocate dataset/model getters, fix imports & global_step
- Rename test_gatv2_pubmed_training.py -> train.py
- Move PubMed loader to blacksmith/datasets/torch/BOUNTIES/pubmed/
- Move model factory to models/ as get_model
- Hoist matplotlib imports to module top; drop CPU-only torch_xla block
- Increment global_step after the optimizer step
