+++ blacksmith/datasets/torch/BOUNTIES/pubmed/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-FileCopyrightText: (c) 2026 Tenstorrent AI ULC
+#
+# SPDX-License-Identifier: Apache-2.0
+from blacksmith.datasets.torch.BOUNTIES.pubmed.pubmed_dataset import load_dataset
+
+__all__ = ["load_dataset"]
